refactor&test: Controller 생성해 서비스를 io로부터 분리하고, 테스트 업데이트
- 원래 Service 내에 있던 private 서비스 메소드들을 하나의 서비스 클래스로 분리
- 원래 IO를 포함했던 테스트에서 io를 제외하고, 서비스만을 테스트
- 테스트 파일 이름 변경
