+++ app/src/test/java/calculator/domain/CalculatorTest.java
@@ -18,7 +18,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class BaseCalculatorTest {
+public class CalculatorTest {
 
     Calculator calculator;
     CalculatorRepository calculatorRepository;
